requirements.txt
google-adk>=1.10.0
python-dotenv>=1.0.1
